Fix boundary conditions for overlap detection
When comparing two annotations to detect overlap, we were failing to
detect some overlaps. If two annotations had identical start times or
identical end times, the overlap of the two wouldn't be caught.

The fix is to use `<=` or `>=` when comparing to `begin` or `end`
values, instead of just `<` or `>`.