Add new EAF summarization script `summarize-eaf.py`
This new version of the counter script does a better job of detecting
overlaps, particularly when more than two tiers are active
simultaneously. `summarize-eaf.py` collects all of the annotated
segments of interest, and generates a list of "events" (times when the
segments start or end) from them, then puts all these events in
chronological order. This technique produces a more accurate union sum
of the amount of time the collective set of tiers is active.

In addition, the output table has been changed to make it more
machine-readable. The columns are now consistent across an entire
corpus, and each row belongs to a unique set of simultaneously active
tiers. It should be possible to use spreadsheet formulas or other
programs on these output tables more easily to do further analyses.

The command-line argument parser has many more options, making it easy
to exclude any tiers from the analysis, change the output field
delimiter, and make the output briefer if overlap details or ADS/CDS
information are not needed.

Last, the script runs ~20 times faster than the old one.