Add tier masking
Now it's possible to use the `--masking-tiers` option to specify a
list of tiers (probably just `CHI`) to use as a mask. Results are
reported as if no tiers are active while any of the specified tiers is
active. This represents a model where the subject (`CHI`) is assumed
to be ignoring input while producing output.