Guarantee that all labels have OutputRecord objects
If a tier (e.g. `TIER1`) consisted solely of segments fully overlapping segments
from other tiers, it would result in no output record being created for that
tier on its own (only `TIER1+TIER2`) in the `output_records` dictionary. When
populating the output record for that tier from the section sums, it would then
try to add an amount to a non-existent entry in the dictionary.

This change guarantees that if a tier appears in an overlap label, it will get
its own separate entry in the `output_records` array, even if all of its
segments are full-overlapping.